# root/back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users": {
      "get": {
        "operationId": "UsersController_findAll",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role used for RBAC. Required for all endpoints except GET /users login lookup.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of all users."
          }
        },
        "summary": "Retrieve all users",
        "tags": [
          "users"
        ]
      },
      "post": {
        "operationId": "UsersController_create",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role used for RBAC. Required for all endpoints except GET /users login lookup.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The user has been successfully created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateUserDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input."
          },
          "403": {
            "description": "Forbidden."
          }
        },
        "summary": "Create a new user (Superuser, Project Manager)",
        "tags": [
          "users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role used for RBAC. Required for all endpoints except GET /users login lookup.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The user details."
          },
          "404": {
            "description": "User not found."
          }
        },
        "summary": "Retrieve a specific user by ID",
        "tags": [
          "users"
        ]
      },
      "patch": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role used for RBAC. Required for all endpoints except GET /users login lookup.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The user has been successfully updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateUserDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "User not found."
          }
        },
        "summary": "Update an existing user (All authenticated roles)",
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role used for RBAC. Required for all endpoints except GET /users login lookup.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The user has been successfully deleted."
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "User not found."
          }
        },
        "summary": "Delete a user (Superuser only)",
        "tags": [
          "users"
        ]
      }
    },
...
    "/messages": {
      "get": {
        "operationId": "MessagesController_findAll",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role used for RBAC. Allowed values depend on the endpoint.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of all messages."
          }
        },
        "summary": "Retrieve all messages",
        "tags": [
          "messages"
        ]
      },
      "post": {
        "operationId": "MessagesController_create",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role used for RBAC. Allowed values depend on the endpoint.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The message has been successfully created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateMessageDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input."
          },
          "403": {
            "description": "Forbidden."
          }
        },
        "summary": "Create a new message",
        "tags": [
          "messages"
        ]
      }
    },
    "/messages/{id}": {
      "get": {
        "operationId": "MessagesController_findOne",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role used for RBAC. Allowed values depend on the endpoint.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The message details."
          },
          "404": {
            "description": "Message not found."
          }
        },
        "summary": "Retrieve a specific message by ID",
        "tags": [
          "messages"
        ]
      },
      "patch": {
        "operationId": "MessagesController_update",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role used for RBAC. Allowed values depend on the endpoint.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMessageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The message has been successfully updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateMessageDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "Message not found."
          }
        },
        "summary": "Update an existing message",
        "tags": [
          "messages"
        ]
      },
      "delete": {
        "operationId": "MessagesController_remove",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role used for RBAC. Allowed values depend on the endpoint.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The message has been successfully deleted."
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "Message not found."
          }
        },
        "summary": "Delete a message",
        "tags": [
          "messages"
        ]
      }
    },
...
  },
  "info": {
    "title": "ConstructHub API",
    "description": "ConstructHub Construction Management Platform API.\n\n**How to use:** Click \"Try it out\" on any endpoint, fill in the `x-role` header field with one of: `superuser`, `project_manager`, `site_engineer`, `finance_manager`, `client`, then click Execute.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Full name of the user",
            "example": "Rajesh Kumar"
          },
          "email": {
            "type": "string",
            "description": "Email address of the user",
            "example": "rajesh@ch.com"
          },
          "password": {
            "type": "string",
            "description": "Password for the user",
            "example": "123456"
          },
          "role": {
            "type": "string",
            "description": "Role of the user",
            "enum": [
              "superuser",
              "project_manager",
              "site_engineer",
              "finance_manager",
              "client"
            ]
          },
          "phone": {
            "type": "string",
            "description": "Phone number of the user",
            "example": "+91 98765 43210"
          },
          "avatar": {
            "type": "string",
            "description": "Avatar initials or URL",
            "example": "RK"
          },
          "status": {
            "type": "string",
            "description": "Status of the user",
            "enum": [
              "active",
              "inactive"
            ],
            "default": "active"
          },
          "createdAt": {
            "type": "string",
            "description": "Date of creation",
            "example": "2026-01-05"
          }
        },
        "required": [
          "name",
          "email",
          "password",
          "role",
          "phone",
          "status"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Full name of the user",
            "example": "Rajesh Kumar"
          },
          "email": {
            "type": "string",
            "description": "Email address of the user",
            "example": "rajesh@ch.com"
          },
          "password": {
            "type": "string",
            "description": "Password for the user",
            "example": "123456"
          },
          "role": {
            "type": "string",
            "description": "Role of the user",
            "enum": [
              "superuser",
              "project_manager",
              "site_engineer",
              "finance_manager",
              "client"
            ]
          },
          "phone": {
            "type": "string",
            "description": "Phone number of the user",
            "example": "+91 98765 43210"
          },
          "avatar": {
            "type": "string",
            "description": "Avatar initials or URL",
            "example": "RK"
          },
          "status": {
            "type": "string",
            "description": "Status of the user",
            "enum": [
              "active",
              "inactive"
            ],
            "default": "active"
          },
          "createdAt": {
            "type": "string",
            "description": "Date of creation",
            "example": "2026-01-05"
          }
        }
      },
...
      "CreateMessageDto": {
        "type": "object",
        "properties": {
          "from": {
            "type": "string",
            "description": "ID of the sender",
            "example": "U002"
          },
          "fromName": {
            "type": "string",
            "description": "Name of the sender",
            "example": "Rajesh Kumar"
          },
          "to": {
            "type": "string",
            "description": "ID of the recipient",
            "example": "U003"
          },
          "toName": {
            "type": "string",
            "description": "Name of the recipient",
            "example": "Priya Sharma"
          },
          "text": {
            "type": "string",
            "description": "Message text content",
            "example": "Please review the updated project timeline."
          },
          "time": {
            "type": "string",
            "description": "Time message was sent",
            "example": "10:30 AM"
          },
          "date": {
            "type": "string",
            "description": "Date message was sent",
            "example": "2026-03-24"
          }
        },
        "required": [
          "from",
          "to",
          "text"
        ]
      },
      "UpdateMessageDto": {
        "type": "object",
        "properties": {
          "from": {
            "type": "string",
            "description": "ID of the sender",
            "example": "U002"
          },
          "fromName": {
            "type": "string",
            "description": "Name of the sender",
            "example": "Rajesh Kumar"
          },
          "to": {
            "type": "string",
            "description": "ID of the recipient",
            "example": "U003"
          },
          "toName": {
            "type": "string",
            "description": "Name of the recipient",
            "example": "Priya Sharma"
          },
          "text": {
            "type": "string",
            "description": "Message text content",
            "example": "Please review the updated project timeline."
          },
          "time": {
            "type": "string",
            "description": "Time message was sent",
            "example": "10:30 AM"
          },
          "date": {
            "type": "string",
            "description": "Date message was sent",
            "example": "2026-03-24"
          }
        }
      },
...
    }
  }
}